Stage of Implementing MecanumDrive..

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/DriveTrain/Mecanum.java
@@ -4,7 +4,7 @@ import org.firstinspires.ftc.teamcode.DriveMotors;
 
 public class Mecanum {
     DriveMotors Motors = new DriveMotors();
-    public void Drive(int xSpeed, int ySpeed, int zRotation) {
+    public void Drive(double xSpeed, double ySpeed, double zRotation) {
         Motors.FLMotor.setPower(xSpeed + ySpeed + zRotation);
         Motors.FRMotor.setPower(xSpeed - ySpeed - zRotation);
         Motors.BLMotor.setPower(xSpeed - ySpeed + zRotation);
